+++ iBMSC/OpVisual.cs
@@ -201,7 +201,7 @@ public partial class OpVisual : Form
                 button20.Size = size;
                 button18.BackColor = col[i].cBG;
                 Button button21 = button18;
-                object obj = Interaction.IIf((int)Math.Round(col[i].cBG.GetBrightness() * 255f) + 255 - unchecked((int)col[i].cBG.A) >= 128, Color.Black, Color.White);
+                object obj = Interaction.IIf((int)Math.Round(col[i].cBG.GetBrightness() * 255f) + 255 - unchecked(col[i].cBG.A) >= 128, Color.Black, Color.White);
                 button21.ForeColor = ((obj != null) ? ((Color)obj) : color);
                 button18.Text = To4Hex(col[i].cBG.ToArgb());
                 button18.Name = "cBG";
@@ -236,7 +236,7 @@ public partial class OpVisual : Form
         xbutton.BackColor = c;
         checked
         {
-            object obj = Interaction.IIf((int)Math.Round(c.GetBrightness() * 255f) + 255 - unchecked((int)c.A) >= 128, Color.Black, Color.White);
+            object obj = Interaction.IIf((int)Math.Round(c.GetBrightness() * 255f) + 255 - unchecked(c.A) >= 128, Color.Black, Color.White);
             Color color = default(Color);
             xbutton.ForeColor = ((obj != null) ? ((Color)obj) : color);
         }
@@ -392,7 +392,7 @@ public partial class OpVisual : Form
                     case "cBG":
                         {
                             button.BackColor = colorPicker.NewColor;
-                            object obj = Interaction.IIf((int)Math.Round(colorPicker.NewColor.GetBrightness() * 255f) + 255 - unchecked((int)colorPicker.NewColor.A) >= 128, Color.Black, Color.White);
+                            object obj = Interaction.IIf((int)Math.Round(colorPicker.NewColor.GetBrightness() * 255f) + 255 - unchecked(colorPicker.NewColor.A) >= 128, Color.Black, Color.White);
                             Color color = default(Color);
                             button.ForeColor = ((obj != null) ? ((Color)obj) : color);
                             break;
